+++ src/__mocks__/global.mock.ts
@@ -1,0 +1,7 @@
+export interface Global {
+  SpreadsheetApp: GoogleAppsScript.Spreadsheet.SpreadsheetApp;
+  PropertiesService: GoogleAppsScript.Properties.PropertiesService;
+  HtmlService: GoogleAppsScript.HTML.HtmlService;
+  UrlFetchApp: GoogleAppsScript.URL_Fetch.UrlFetchApp;
+  Logger: GoogleAppsScript.Base.Logger;
+}
