--- src/helpers/html.module.ts
@@ -1,7 +1,0 @@
-/*
-* Wildcard module which allows importing HTML to TS files.
-*/
-declare module "*.html" {
-  const value: string;
-  export default value;
-}
