--- src/interfaces/google-helpers.ts
@@ -1,7 +1,0 @@
-/**
- * Simple helper interface which saves typing Array<Array<number | boolean | string | Date>> for range.GetValues().
- * @interface IGetValuesResult
- * @extends {(Array<Array<number | boolean | string | Date>>)}
- */
-interface IGetValuesResult
-  extends Array<Array<number | boolean | string | Date>> {}
